+++ addOns/scripts/src/main/java/org/zaproxy/zap/extension/scripts/scanrules/ActiveScriptScanRule.java
@@ -237,6 +237,14 @@ public class ActiveScriptScanRule extends ActiveScriptHelper {
         return script.isEnabled();
     }
 
+    public String getCodeLink() {
+        return metadata.getCodeLink();
+    }
+
+    public String getHelpLink() {
+        return metadata.getHelpLink();
+    }
+
     private ExtensionScript getExtScript() {
         if (extScript == null) {
             extScript =