+++ addOns/scripts/src/main/java/org/zaproxy/zap/extension/scripts/scanrules/PassiveScriptScanRule.java
@@ -120,6 +120,14 @@ public class PassiveScriptScanRule extends PassiveScriptHelper {
         return script.isEnabled();
     }
 
+    public String getCodeLink() {
+        return metadata.getCodeLink();
+    }
+
+    public String getHelpLink() {
+        return metadata.getHelpLink();
+    }
+
     @Override
     public List<Alert> getExampleAlerts() {
         return List.of(newAlert().build());