+++ addOns/exim/src/main/java/org/zaproxy/addon/exim/AbstractPopupMenuSaveMessage.java
@@ -95,11 +95,6 @@ abstract class AbstractPopupMenuSaveMessage extends PopupMenuHttpMessageContaine
     }
 
     @Override
-    public boolean precedeWithSeparator() {
-        return true;
-    }
-
-    @Override
     public boolean isSafe() {
         return true;
     }