+++ addOns/spider/src/main/java/org/zaproxy/addon/spider/PopupMenuItemSpiderDialogWithContext.java
@@ -48,6 +48,7 @@ public class PopupMenuItemSpiderDialogWithContext extends PopupContextTreeMenu {
                 });
     }
 
+    @Override
     public int getWeight() {
         return MenuWeights.MENU_CONTEXT_SPIDER_WEIGHT;
     }