+++ addOns/spiderAjax/src/main/java/org/zaproxy/zap/extension/spiderAjax/PopupMenuAjaxSite.java
@@ -49,16 +49,12 @@ public class PopupMenuAjaxSite extends PopupMenuItemSiteNodeContainer {
         return Constant.messages.getString("attack.site.popup");
     }
 
-    /** */
     @Override
-    public int getParentMenuIndex() {
-        return ATTACK_MENU_INDEX;
-    }
-
     public int getParentWeight() {
         return MenuWeights.MENU_ATTACK_WEIGHT;
     }
 
+    @Override
     public int getWeight() {
         return MenuWeights.MENU_ATTACK_AJAX_WEIGHT;
     }