+++ addOns/zest/src/main/java/org/zaproxy/zap/extension/zest/menu/ZestAddActionPopupMenu.java
@@ -177,7 +177,6 @@ public class ZestAddActionPopupMenu extends ExtensionPopupMenuItem {
                                 .showZestActionDialog(parent, child, stmt, za, true);
                     }
                 });
-        menu.setMenuIndex(this.getMenuIndex());
         View.getSingleton().getPopupList().add(menu);
     }