+++ addOns/zest/src/main/java/org/zaproxy/zap/extension/zest/menu/ZestRecordFromNodePopupMenu.java
@@ -48,6 +48,7 @@ public class ZestRecordFromNodePopupMenu extends PopupMenuItemSiteNodeContainer 
         extension.getDialogManager().showZestRecordScriptDialog(node);
     }
 
+    @Override
     public int getWeight() {
         return MenuWeights.MENU_SCRIPT_ZEST_RECORD_WEIGHT;
     }