+++ addOns/scripts/CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+
 ## [45.2.0] - 2024-04-11
 ### Added
 - Active and Passive Scripts with a `getMetadata()` function are now treated as first-class scan rules (Issue 7105).