+++ addOns/scripts/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support for code and help links for script scan rules.
 
 ### Changed
+- Update minimum ZAP version to 2.15.0.
 - Allow to set raw parameter values from Active Rules, by calling `as.setEscapedParam(HttpMessage msg, String param, String value)`.
 
 ## [45.2.0] - 2024-04-11